Added signup endpoint for AppUsers with non-guest role to signup. (#95)
* Signup endpoint is done, just need to test now.

* Added block to prevent guests from signing up.

* Turned SignupService into a class, changed implementation of isGuest().

Also added docs to the methods w/o docs in SignupService and
AppUserService.

* Changed name of SignupService to AccountService.
