+++ src/services/index.ts
@@ -4,3 +4,4 @@ export { AttendanceService, AttendanceServiceImpl } from './AttendanceService';
 export { RSVPService, RSVPServiceImpl } from './RSVPService';
 export { AuthenticationService, AuthenticationServiceImpl } from './AuthenticationService';
 export { AuthorizationService, AuthorizationServiceImpl } from './AuthorizationService';
+export { AccountService, AccountServiceImpl } from './AccountService';
